style: add comments to session classes
