--- plugin-server/src/main/ingestion-queues/session-recording-v2/sessions/session-batch-writer.ts
@@ -1,10 +1,0 @@
-import { Writable } from 'stream'
-
-export interface StreamWithFinish {
-    stream: Writable
-    finish: () => Promise<void>
-}
-
-export interface SessionBatchWriter {
-    newBatch(): Promise<StreamWithFinish>
-}
